pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["adrian <mitomono@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
strawberry-graphql = { extras = ["fastapi"], version = "^0.215.0" }
psycopg2 = "^2.9.9"
click = "^8.1.7"
graphql-core = "^3.2.3"
h11 = "^0.14.0"
httptools = "^0.6.1"
idna = "^3.4"
ruff = "^0.1.6"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.1"
python-dateutil = "^2.8.2"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
pyyaml = "^6.0.1"
six = "^1.16.0"
sniffio = "^1.3.0"
typing-extensions = "^4.8.0"
watchfiles = "^0.21.0"
websockets = "^12.0"
orjson = "^3.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
